installer script changed to support remote installation.

+++ install.sh
@@ -1,11 +1,17 @@
 #! /bin/bash
 
-VERSION=$(emacs -l src/lib/version.el --eval "(princ (get_version))" --batch -Q)
+VERSION=1.0.0
+
+remoteconffile=
+remoteexecutable=
+
+conffile=conf/dotkuso
 
 # Gathering informations
-echo -e "\n\tKuso IDE $VERSION copyright 2010-2011 Sameer Rahmani <lxsameer@gnu.org>\n\n"
+echo -e "\n\033[01;32mKuso IDE\033[00m $VERSION copyright 2010-2013 \033[01;34mSameer Rahmani <lxsameer@gnu.org>\033[00m\n\n"
 echo "Enter requested informations. You can change it later in top"
-echo -e "level customization.\n\n"
+echo -e "your init file.\n\n"
+
 condition="1"
 while [ "$condition" == "1" ] ; do
     read -p "Do you want to install Kuso IDE as an stand alone application ([y]/n)? " standalone
@@ -13,10 +19,10 @@ while [ "$condition" == "1" ] ; do
     if [ "$standalone" == "" -o "$standalone" == "y" ]
     then
 	standalone="y"
-	dotemacs=~/.kuso
-	repo=~/.kuso.d
+	dotemacs=~/.kuso_dev
+	repo=~/.kuso.d_dev
 	condition="0"
-	conffile=conf/dotkuso
+	executable=kuso-dev
     fi
 
     if [ "$standalone" == "n" ]
@@ -24,9 +30,9 @@ while [ "$condition" == "1" ] ; do
 	dotemacs=~/.emacs
 	repo=~/.emacs.d
 	condition="0"
-	conffile=conf/dotemacs
+
+	executable=emacs-dev
     fi
-    
 
 done
 read -p "Enter your full name: " fullname
@@ -45,35 +51,32 @@ addr=$HOME/.kuso.d
 kusohome=`pwd`
 mkdir -p $repo
 if [ -e $dotemacs ]; then
-    echo "Backing up exists .emacs file . . ."
-    cp $dotemacs "$dotemacs.backup"
+    echo "Backing up exists init file . . ."
+    #cp $dotemacs "$dotemacs.backup"
 fi
 echo "Copying files . . . "
-cp conf/emacs.d/* $repo -r
-mkdir -p $addr
-mkdir -p $HOME/.tmp
-cp conf/bin/pyemacs.sh $addr/ -r
-chmod u+x $addr/pyemacs.sh
-
-echo "Creating ~/.emacs"
-cp $conffile $dotemacs
-
-if [ "$standalone" == "y" ]
+if [ -e $conffile ]
 then
-    cp bin/kuso $repo
-    cp bin/kuso.desktop $HOME/.local/share/applications/
-    sed "s,--HOME--,$HOME,mg" -i $HOME/.local/share/applications/kuso.desktop
-    cp images/icon.svg $repo
+    cp $conffile $dotemacs
+    cp bin/$executable $repo/$executable
+else
+    wget $remoteconffile -o $dotemacs
+    wget $remoteexecutable -o $repo/$executable
+fi
+
+if [ "$standalone" == "" -o "$standalone" == "y" ]
+then
+    sudo ln -s $repo/$executable /usr/bin/$executable
 fi
 
 sed "s/--EMAIL--/$mail/mg" -i $dotemacs
 sed "s/--FULLNAME--/$fullname/mg" -i $dotemacs
 sed "s,--WORKSPACE--,$workspace,mg" -i $dotemacs
-sed "s,--ADDR--,$addr,mg" -i $dotemacs
-sed "s,--KUSOHOME--,$kusohome,mg" -i $dotemacs
+sed "s,--REPO--,$repo,mg" -i $dotemacs
+
 
 echo "Copy the below code in your initial shell script:"
-echo 
+echo
 echo "export PATH=\$PATH:$repo"
 echo -e "\nInstallation finished."
 echo "Restart the GNU/Emacs and make sure that all the requirements met."
